refactor routes to match titles
