+++ wildmile/app/api/cameratrap/analytics/observation-activity/route.js
